getFloat bug fix

+++ app/src/main/java/com/example/recipeapp/HomeFragment.java
@@ -418,7 +418,7 @@ public class HomeFragment extends Fragment {
             cursor1.moveToNext();
             String itemCountString = String.valueOf(cursor1.getInt(0)) + " items";
             String placeCountString = "(" + String.valueOf(cursor1.getInt(1)) + " places)";
-            String costString = String.valueOf(cursor1.getInt(2));
+            String costString = String.valueOf(cursor1.getFloat(2));
 
             lastShoppingListItemCount.setText(itemCountString);
             lastShoppingListPlaceCount.setText(placeCountString);
